refacto: moving CreateBankingStatusBox off DaggerfallUI to lower bloat
That MessageBox does not have the same lifecycle as
DaggerfallBankingWindow, so I converted it to static method but this is
not very satisfying either.
